services/post_streaming/pyproject.toml
[tool.poetry]
name = "src"
version = "0.0.1"
description = "Siftino"
authors = [
    "Jonas Lindberg <jonas.lindberg@futi.no>",
    "Jorge Lewis <jorge.lewis@futi.no>",
]
readme = "./README.md"
homepage = "https://relevanti.no"
repository = "https://github.com/startino/relevantino"
documentation = "https://api.rest.futi.no/docs"
license = "BSL-1.1 (https://github.com/startino/relevantino/blob/alpha/LICENSE)"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.27.1"
fastapi = "^0.110.0"
tenacity = "^8.2.3"
supabase = "^2.4.0"
aiohttp = "^3.9.4"
redis = "^5.0.2"
langchain = "^0.1.12"
langchain-community = "^0.0.38"
langchain-openai = "^0.1.1"
praw = "^7.7.1"
markdown = "^3.6"
gptrim = "^0.1.6"
pandas = "^2.2.1"
reddit-chatbot-python = "^1.4.9"
selenium = "^4.19.0"
mail = "^2.1.0"
python-dotenv = "^1.0.1"
diskcache = "^5.6.3"
pydantic = "^2.6.4"
python-upwork-oauth2 = "^3.1.0"
sse-starlette = "^2.1.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.7.0"
black = "^24.3.0"
isort = "^5.12.0"
autoflake = "^2.2.1"
flake8 = "^7.0.0"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
pyright = "^1.1.352"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
